Switch to angular-cli@webpack
The newest beta from angular-cli simplifies lots of configurations,
relying on Webpack rather that SystemJS to build all files.
The lint phase is now included in tests, which led to the discovery
of several syntax warnings.
Tests are working with a deprecated method (addProvider), which
will have to be replaced before upgrading to the next angular-cli release.